Added: Added Order id,date,user,book and quantity details to book store

+++ src/main/java/com/bridgelabz/bookstore/dto/CartDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class CartDTO {
     long userId;
-    List<Integer> bookIdList;
-    List<Integer> quantity;
+    List<Long> bookIdList;
+    List<Long> quantity;
 }
